Added Balloons and DEXTemplate contracts and deploy script
