--- packages/hardhat-ts/contracts/ExampleExternalContract.sol
@@ -1,12 +1,0 @@
-pragma solidity >=0.8.0 <0.9.0;
-
-//SPDX-License-Identifier: MIT
-
-contract ExampleExternalContract {
-
-    bool public completed;
-
-    function complete() public payable {
-        completed = true;
-    }
-}
