--- packages/hardhat-ts/contracts/Staker.sol
@@ -1,25 +1,0 @@
-pragma solidity >=0.8.0 <0.9.0;
-//SPDX-License-Identifier: MIT
-
-import 'hardhat/console.sol';
-import './ExampleExternalContract.sol';
-
-contract Staker {
-  ExampleExternalContract public exampleExternalContract;
-
-  constructor(address exampleExternalContractAddress) public {
-    exampleExternalContract = ExampleExternalContract(exampleExternalContractAddress);
-  }
-
-  // TODO: Collect funds in a payable `stake()` function and track individual `balances` with a mapping:
-  //  ( make sure to add a `Stake(address,uint256)` event and emit it for the frontend <List/> display )
-
-  // TODO: After some `deadline` allow anyone to call an `execute()` function
-  //  It should call `exampleExternalContract.complete{value: address(this).balance}()` to send all the value
-
-  // TODO: if the `threshold` was not met, allow everyone to call a `withdraw()` function
-
-  // TODO: Add a `timeLeft()` view function that returns the time left before the deadline for the frontend
-
-  // TODO: Add the `receive()` special function that receives eth and calls stake()
-}
