--- packages/hardhat-ts/deploy/00_deploy_example_external_contract.ts
@@ -1,33 +1,0 @@
-import { DeployFunction } from 'hardhat-deploy/types';
-import { HardhatRuntimeEnvironmentExtended } from 'helpers/types/hardhat-type-extensions';
-
-const func: DeployFunction = async (hre: HardhatRuntimeEnvironmentExtended) => {
-  const { getNamedAccounts, deployments } = hre as any;
-  const { deploy } = deployments;
-  const { deployer } = await getNamedAccounts();
-  await deploy('ExampleExternalContract', {
-    // Learn more about args here: https://www.npmjs.com/package/hardhat-deploy#deploymentsdeploy
-    from: deployer,
-    // args: ["Hello"],
-    log: true,
-  });
-
-  /*
-    // Getting a previously deployed contract
-    const YourContract = await ethers.getContract("YourContract", deployer);
-    await YourContract.setPurpose("Hello");
-
-    //const yourContract = await ethers.getContractAt('YourContract', "0xaAC799eC2d00C013f1F11c37E654e59B0429DF6A") //<-- if you want to instantiate a version of a contract at a specific address!
-  */
-};
-export default func;
-func.tags = ['ExampleExternalContract'];
-
-/*
-Tenderly verification
-let verification = await tenderly.verify({
-  name: contractName,
-  address: contractAddress,
-  network: targetNetwork,
-});
-*/
